+++ app/frontend/src/services/projects.ts
@@ -1,6 +1,6 @@
 import { Project } from 'types/entities';
 import { CreateProjectResponsePayload } from 'types/transport/responses/project/create';
-import client from 'utils/ctproto-client';
+import client from 'transport/ctproto-client';
 
 /**
  * Creates new project
